fix log initalising
