+++ inject/wire_gen.go
@@ -29,8 +29,6 @@ func Inject() file.FileProcessor {
 var FileProcessor = Inject()
 
 func ConfigSetup() config.Config {
-	viper.AutomaticEnv()
-	configureLogging()
 	viper.SetDefault("PORT", "8080")
 	viper.SetDefault("PROJECT_ID", "rm-ras-sandbox")
 	viper.SetDefault("TOPIC_ID", "topic")
@@ -46,16 +44,6 @@ func ConfigSetup() config.Config {
 		},
 	}
 	return config2
-}
-
-func configureLogging() {
-	verbose := viper.GetBool("VERBOSE")
-	logrus.SetFormatter(&logrus.JSONFormatter{})
-	if verbose {
-		logrus.SetLevel(logrus.DebugLevel)
-	} else {
-		logrus.SetLevel(logrus.InfoLevel)
-	}
 }
 
 func NewFileProcessor(config2 config.Config, client *pubsub.Client, ctx context.Context) file.FileProcessor {
